Move GetAddress to Location class
