+++ src/Net/Models/GymDetailsData.cs
@@ -2,11 +2,10 @@
 {
     using System;
     using System.Collections.Generic;
-    using System.IO;
     using System.Linq;
+
     using DSharpPlus;
     using DSharpPlus.Entities;
-
     using Newtonsoft.Json;
     using ServiceStack.DataAnnotations;
     using ServiceStack.OrmLite;
@@ -16,6 +15,7 @@
     using WhMgr.Configuration;
     using WhMgr.Data;
     using WhMgr.Diagnostics;
+    using WhMgr.Geofence;
     using WhMgr.Utilities;
 
     /// <summary>
@@ -145,7 +145,7 @@
             var appleMapsLocationLink = UrlShortener.CreateShortUrl(whConfig.ShortUrlApiUrl, appleMapsLink);
             var wazeMapsLocationLink = UrlShortener.CreateShortUrl(whConfig.ShortUrlApiUrl, wazeMapsLink);
             var scannerMapsLocationLink = UrlShortener.CreateShortUrl(whConfig.ShortUrlApiUrl, scannerMapsLink);
-            var address = Utils.GetAddress(city, Latitude, Longitude, whConfig);
+            var address = new Location(null, city, Latitude, Longitude).GetAddress(whConfig);
             //var staticMapLocationLink = string.IsNullOrEmpty(whConfig.ShortUrlApiUrl) ? staticMapLink : NetUtil.CreateShortUrl(whConfig.ShortUrlApiUrl, staticMapLink);
 
             const string defaultMissingValue = "?";
