+++ src/Utilities/Utils.cs
@@ -66,70 +66,6 @@
             return message.ErrorCode == null;
         }
 
-        public static Location GetAddress(string city, double lat, double lng, WhConfig config)
-        {
-            if (!string.IsNullOrEmpty(config.GoogleMapsKey))
-                return GetGoogleAddress(city, lat, lng, config.GoogleMapsKey);
-
-            if (!string.IsNullOrEmpty(config.NominatimEndpoint))
-                return GetNominatimAddress(city, lat, lng, config.NominatimEndpoint);
-
-            return null;
-        }
-
-        public static Location GetGoogleAddress(string city, double lat, double lng, string gmapsKey)
-        {
-            var apiKey = string.IsNullOrEmpty(gmapsKey) ? string.Empty : $"&key={gmapsKey}";
-            var url = $"https://maps.googleapis.com/maps/api/geocode/json?latlng={lat},{lng}&sensor=true{apiKey}";
-            var unknown = "Unknown";
-            try
-            {
-                var request = (HttpWebRequest)WebRequest.Create(url);
-                var response = request.GetResponse();
-                using (var responseStream = response.GetResponseStream())
-                {
-                    var reader = new StreamReader(responseStream, Encoding.UTF8);
-                    var data = reader.ReadToEnd();
-                    var parseJson = JObject.Parse(data);
-                    var status = Convert.ToString(parseJson["status"]);
-                    if (string.Compare(status, "OK", true) != 0)
-                        return null;
-
-                    var result = parseJson["results"].FirstOrDefault();
-                    var address = Convert.ToString(result["formatted_address"]);
-                    //var area = Convert.ToString(result["address_components"][2]["long_name"]);
-                    return new Location(address, city ?? unknown, lat, lng);
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex);
-            }
-            return null;
-        }
-
-        public static Location GetNominatimAddress(string city, double lat, double lng, string endpoint)
-        {
-            var unknown = "Unknown";
-            var url = $"{endpoint}/reverse?format=jsonv2&lat={lat}&lon={lng}";
-            try
-            {
-                using (var wc = new WebClient())
-                {
-                    wc.Proxy = null;
-                    wc.Headers.Add("User-Agent", Strings.BotName);
-                    var json = wc.DownloadString(url);
-                    dynamic obj = JsonConvert.DeserializeObject(json);
-                    return new Location(Convert.ToString(obj.display_name), city ?? unknown, Convert.ToDouble(obj.lat), Convert.ToDouble(obj.lon));
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.Error(ex);
-            }
-            return null;
-        }
-
         public static double GetUnixTimestamp()
         {
             return DateTime.Now.Subtract(new DateTime(1970, 1, 1)).TotalSeconds;
